+++ backup/mysql/demo.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-mysql -uroot -p${MYSQL_ROOT_PASSWORD} < /backup/demo.sql